requirements.txt
nsepython==2.4
supabase==2.3.4
python-dotenv==1.0.0
pytz==2024.1
